fix bugs, retune eval, refactor code ``` Score of New vs Old: 278 - 244 - 255  [0.522] 777 ...      New playing White: 151 - 123 - 115  [0.536] 389 ...      New playing Black: 127 - 121 - 140  [0.508] 388 ...      White vs Black: 272 - 250 - 255  [0.514] 777 Elo difference: 15.2 +/- 20.0, LOS: 93.2 %, DrawRatio: 32.8 % SPRT: llr 0.609 (20.7%), lbound -2.94, ubound 2.94 ```

+++ src/search.cpp
@@ -39,7 +39,7 @@ static void checkLimit(SearchState& ss) {
     if (ss.limits.nodes && ss.nodes >= (uint64_t)ss.limits.nodes)
         ss.stop = true;
     else if (ss.timeLimit) {
-        if (elapsedMs(ss) >= ss.timeLimit)
+        if ((int)elapsedMs(ss) >= ss.timeLimit)
             ss.stop = true;
     }
 }
@@ -304,7 +304,7 @@ SearchResult search(SearchState& ss, Board& board, const Limits& limits,
     Move prevBest = MOVE_NONE;
     for (int depth = 1; depth <= maxDepth && !ss.stop; ++depth) {
         if (depth > 1 && ss.softTime) {
-            uint64_t elapsed = elapsedMs(ss);
+            int elapsed = (int)elapsedMs(ss);
             if (elapsed >= ss.timeLimit) break;
             if (elapsed > ss.softTime && stable) break;
         }
